re-factors tests into separate files
